# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel Swagger",
        "version": "1.0.0"
    },
    "paths": {
        "/api/movies": {},
        "/api/": {},
        "/api/client/movies": {
            "get": {
                "tags": [
                    "movies"
                ],
                "summary": "Получение списка фильмов",
                "operationId": "f7ef45dda4b579c0e5c9343350e8d038",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Номер страницы",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список фильмов",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/MovieResource"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "description": "Ошибка",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/client/movies/{id}": {
            "get": {
                "tags": [
                    "movies"
                ],
                "summary": "Получение информации о фильме",
                "operationId": "e92c32f136c14b3237c0ab46a91f8dfe",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID фильма",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Информация о фильме",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MovieResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Фильм не найден",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "description": "Ошибка",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "description": "Ошибка",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/client/movies/{id}/tickets": {
            "get": {
                "tags": [
                    "movies"
                ],
                "summary": "Получение билетов для фильма",
                "operationId": "08df6b7bafcf4761f05c19bd337d2ef5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID фильма",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список билетов",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/TicketResource"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Фильм не найден",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "description": "Ошибка",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "description": "Ошибка",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "MovieResource": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "title": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "release_date": {
                        "type": "string",
                        "format": "date"
                    },
                    "duration": {
                        "type": "integer"
                    },
                    "rating": {
                        "type": "number",
                        "format": "float"
                    }
                },
                "type": "object"
            },
            "TicketResource": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "movie_id": {
                        "type": "integer"
                    },
                    "price": {
                        "type": "number",
                        "format": "float"
                    },
                    "available": {
                        "type": "boolean"
                    },
                    "showtime": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "movies",
            "description": "movies"
        }
    ]
}